% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_WES_standard_curve.R
\docType{data}
\name{template_WES_standard_curve}
\alias{template_WES_standard_curve}
\title{Template standard curve data}
\format{
\subsection{\code{template_WES_standard_curve}}{

A data frame with 3 columns:
\describe{
\item{target_name}{The unique name of the gene target for which the Ct values correspond.}
\item{n_copies}{The number of gene copies represented in the particular dilution.}
\item{ct_value}{The Cycle Threshold (Ct) of the qPCR assay.}
}
}
}
\usage{
template_WES_standard_curve
}
\description{
The \code{template_WES_standard_curve} object provides a template of the data format required
by the 'WES' package for standard curve values. These data are only required when calculating the
number of gene copies using the \code{calc_n_copies} function.
}
\keyword{datasets}
